Copy working LList into ScalaJsonSpec, write failing nested HList format
