+++ support/scalajson/src/test/scala/sjsonnew/support/scalajson/unsafe/HListFormatSpec.scala
@@ -9,17 +9,17 @@ import org.scalatest._
 class HListFormatSpec extends FlatSpec {
   case class Peep(name: String, age: Int)
 
-  import LList.:*:
-  type PeepRepr = String :*: Int :*: LNil
-
-  implicit val PeepIso: IsoLList.Aux[Peep, PeepRepr] = LList.iso(
-    { p: Peep => ("name", p.name) :*: ("age", p.age) :*: LNil },
-    { in: PeepRepr => Peep(in.head, in.tail.head) }
-  )
+  import HList.:+:
+  implicit val PeepFormat: JsonFormat[Peep] = {
+    project[Peep, String :+: Int :+: HNil](
+      p => p.name :+: p.age :+: (HNil: HNil),
+      { case name :+: age :+: HNil => Peep(name, age) }
+    )
+  }
 
   val bob = Peep("Bob", 23)
-  val bobJson = JObject(JField("name", JString("Bob")), JField("age", JNumber(23)))
-  val bobJsonStr = """{"name":"Bob","age":23}"""
+  val bobJson = JArray(JString("Bob"), JNumber(23))
+  val bobJsonStr = """["Bob",23]"""
 
   it should "Peep.toJson correctly"              in assert(bob.toJson === bobJson)
   it should "Peep.toJsonStr correctly"           in assert(bob.toJsonStr === bobJsonStr)
@@ -27,7 +27,7 @@ class HListFormatSpec extends FlatSpec {
   it should "String.fromJsonStr[Peep] correctly" in assert(bobJsonStr.fromJsonStr[Peep] === bob)
   it should "Peep.jsonRoundTrip correctly"       in assertRoundTrip(bob)
 
-  it should "HList.jsonRoundTrip correctly" in assertRoundTrip(bob :+: Peep("Amy", 22) :+: HNil)
+  it should "HList.jsonRoundTrip correctly" in pendingUntilFixed(assertRoundTrip(bob :+: Peep("Amy", 22) :+: HNil))
 
   private def assertRoundTrip[A: JsonWriter : JsonReader](x: A) = assert(x === x.jsonRoundTrip)
 }
